// src/graphql.schema.ts
export class CustomerStats {
    total?: number;
    totalActive?: number;
    totalInactive?: number;
}